+++ packages/toprf-secure-backup/src/index.ts
@@ -5,9 +5,6 @@ export { TOPRFError, TOPRFErrorCode } from './errors';
 export * from './utils';
 export { EncAccountDataType } from './constants';
 export type {
-  SecretDataItem,
-  SecretDataItemInput,
-  SecretDataItemOutput,
   UpdateSecretDataItem,
   MetadataUpdateSecretDataItemParams,
   MetadataBatchUpdateSecretDataItemParams,